chore: fix lint error and update server script
