+++ client/src/components/ModelDetailPage.tsx
