+++ client/src/components/PredictionInterface.tsx
