+++ client/src/components/ProjectDetail.tsx
