+++ client/src/components/ResultsDashboard.tsx
