+++ client/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import App from "../src/App.tsx";
-import './index.css';
+import App from "./App";
+import '../index.css';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
