+++ server/package.json
@@ -4,7 +4,7 @@
   "type": "module",
   "scripts": {
     "dev": "tsx server.ts",
-    "start": "node server.js",
+    "start": "node dist/server.js",
     "build": "tsc"
   },
   "dependencies": {
